Tasking: Introduce ConcurrentCall
It substitutes Utils::Async, but is Utils independent.

The ConcurrentCall class is prepared to work with TaskTree.
Provide the task tree adapter for the ConcurrentCall class.
Register the task inside the Tasking namespace under
the ConcurrentCallTask name.

This class introduces the dependency to Qt::Concurrent,
otherwise Tasking namespace is independent on Qt::Concurrent.
Possibly, may be added into Qt::Concurrent lib,
as a wrapper around the QtConcurrent::run() call.

Change-Id: I4511ff0430e78346aa6a4fae1a9d5370fdd08506
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Marcus Tillmanns <marcus.tillmanns@qt.io>
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>