Debugger test: Automatically sign on macOS
The "gui" manual test is used for release testing attaching a running
application and a core file to the debugger. For this, the application
must be signed with the right entitlements.

Since ad-hoc signing is sufficient, this doesn't require any setup on the
testers computer. So, make this signing automatic.

Change-Id: I7042a3091848ee1c9f7f3843507f3ca616d36179
Reviewed-by: Robert Löhning <robert.loehning@qt.io>
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>
Reviewed-by: Christian Stenger <christian.stenger@qt.io>