RemoteLinux: Generalize deploy step
Instead of three different deploy steps we use only the RSyncDeployStep
which can be configured to each of the three file transfer methods.

The SupportsRSync constant is moved to projectexplorer so that it can be
used by all devices.

The RSyncDeployStep checks which transfer method is available based on
source and target devices. An option is added for the user to force
a different method in case of issues.

Change-Id: I2b1eb39b5bd4a30d0f879d18317b7677d127f48f
Reviewed-by: hjk <hjk@qt.io>