Qbs build: Fix app_version generation
Amends 6836caa14c247829bfd4edf47d31c87310d56211

which added a variable to app_version.h.cmake, but not to
app_version.h.in. Fix by porting qbs to use app_version.h.cmake, which
is the truth nowadays, instead of the app_version.h.in which originated
from the qmake build.

Change-Id: Ia9be53377a0bfded91ccd6ea1637eeda7fe9d5c8
Reviewed-by: Christian Stenger <christian.stenger@qt.io>