Disable output view actions if they are not applicable
In Qt Creator the output views are always accessible: The status bar
shows their buttons, and if an output view is opened in a mode that
doesn't have an OutputPanePlaceHolder (which defines the location), the
edit mode is opened.

This is not the case in Design Studio: The status bar is different, and
design mode can only be opened with an open project, so e.g. after
startup in the welcome screen the output view actions should be
disabled.

Achieve that by disabling the actions when the output view buttons in
the status bar are not shown and the current mode doesn't have an
OutputPanePlaceHolder.

Amends 4642c5fd7ddf73b4753037c3616c2d60d5899b6e
and 3befa60970f2dac2223df4f017c74f7444596015

Change-Id: I57b9ce3f80006fc2e842e38e93f1f7b50c5749f1
Reviewed-by: Henning Gründl <henning.gruendl@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>