Core: Introduce Menu and Action classes
Idea is to have a bit more "object { propertySetter }" kind of pattern
for common cases and to couple actions and menus only via Ids, not
real objects.

In a LayoutBuilder context this might later look like

    Menu {
        id(BOOKMARKS_MENU),
        title(Tr::tr("&Bookmarks")),
        container(Core::Constants::M_TOOLS)
    }

in the end.

Overall line count increases but I claim it reduces complexity on the
user code side.

Change-Id: I166e0c8c6b3036430dc464fd652fb37ded21c09a
Reviewed-by: Eike Ziller <eike.ziller@qt.io>