Wizards: Improve the bundle id for mac/iOS
If an iOS developer doesn't have a wildcard app identifier registered,
the app identifier must be in some sense "unique". So just
"my.example.com" creates issues in this situation.

Qt >= 6.1 tries to set a better bundle id, reading Xcode settings for
the last used prefix there, and Qt 6.4 improves on that in case this
Xcode setting is not available. So, try to use this mechanism in the
wizards.

Even with the automatic deduction of a bundle id, this might be
completely wrong (the Xcode setting is whatever was used last in a Xcode
wizard, regardless of target platform, it can be different for different
users, or for different teams that the user works in), and Qt doesn't
set it for macOS, so aside from out-of-the-box experience the user
should set an explicit bundle at some point.

So, keep the CMake property mentioned with a comment in any case.

Fixes: QTCREATORBUG-29340
Change-Id: I9dacae5c693ec2e23a9cf8bad660ad939303dbf3
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>