Terminal: Fix process staying alive
When starting a Terminal Process with TerminalMode::Detached,
the Process object was not deleted on app exit and therefor
created warnings on exit.

This patch changes it so that the osascript does not wait for the
process to exit and therefor exits pretty much immediately which
deletes the Process object.

Fixes: QTCREATORBUG-29375
Change-Id: I08e8b753c98011fdafc32f03886b5c0be7aec801
Reviewed-by: Eike Ziller <eike.ziller@qt.io>
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>