Fix button states at startup
When opening Qt Creator with clean settings, the run, debug and build
actions build be in the wrong state (enabled instead of disabled, no
debug icon).

This happened because a ProjectExplorer::updateActions call was removed
from the end of restoring the session at startup when moving the session
handling to Core plugin.

That updateActions call shouldn't be necesseray, and isn't if a session
is loaded, even if it is the "implicit default", because then
sessionLoaded(...) is emitted which results in the updateActions() call.

The sessionLoaded(...) signal was not sent if the file for the default
session does not exist. Fix that by also sending the signal even if no
file was actually loaded.

Amends cbeac777a61322fbc2d365c7b83ead8aaaaf4bbf

Change-Id: Idb625253f29611e777d85f0feb6c196fae567ed0
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>