Vcpkg: Implement adding of dependencies to vcpkg.json
Instead of just inserting the package name at the cursor position, Qt
Creator now attempts to append the package name in the dependencies
array.

With auto tests.

The "interesting" effect of this is that the order of values inside
a Json object may change. Qt seems to enforce an alphabetical order
on the key names.

Also the indentation becomes 4 after the modification.

Task-number: QTCREATORBUG-29333
Change-Id: I803ca6a19ee32aeba95640726ac8d8e477750bc0
Reviewed-by: Cristian Adam <cristian.adam@qt.io>