Utils: Fix inverted logic for checkableDecider (continued)
Amends 6b6b1198dd5050a7a896ac9c7505ebcf2c9bce17

The commit above made sure that the display of the dialog was in sync
with the checkbox in the settings dialog.

Now also the "Do not show again" checkbox in the messagebox will be
reflected in the settings dialog.

Change-Id: Ieb186b134adac09c9846a64144056b3b89d297d8
Reviewed-by: hjk <hjk@qt.io>