Issues: Prevent ever-growing session
The session contains a list of disabled issue categories, and that was
growing with each start of Qt Creator, accumulating entries for
categories that are added as "disabled by default".

When adding a category as "disabled by default", it would
unconditionally append it to the list of disabled categories. This list
is saved and restored at Qt Creator restart, and then, when the category
is added by the code as disabled by default, it was appended again, now
appearing twice.

Use a set for the disabled categories internally to avoid this.

Change-Id: I0e2fae92d5b78d6bcc13d946f27241ddb8da84b6
Reviewed-by: Christian Stenger <christian.stenger@qt.io>