Editors: Fix that history of editors could grow
The "history of editors" is supposed to contain the order in which files/
documents were recently accessed. It is used for the "open documents
window" (the Ctrl+Tab popup). In contrast to the "navigation history" it
should contain each file/document only once.

Since 324de13b4e6703db778010d0682ac26cff359516 items for suspended
documents with the same file name were no longer removed. For suspended
documents it is necessary to check for the same file name.

This was so far not vital, but when we save the editor history in the
session, these do not have a document when restoring, which would lead
to an evergrowing editor history.

Amends 324de13b4e6703db778010d0682ac26cff359516

Change-Id: Ia4b7848a1265024d0463afbf7c1cd69189c4be97
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Christian Stenger <christian.stenger@qt.io>
Reviewed-by: David Schulz <david.schulz@qt.io>