Designer: Add command line arguments for plugin paths
Add a command line argument for overriding the default plugin path
(useful if you have installed custom designer plugins in a different,
compatible Qt version), and one for adding plugin paths (for example to
separate builds of custom designer plugins that haven't been installed
anywhere).

This feature is only enabled when building with Qt >= 6.7, which added
the option to change the plugin paths for designer.

Change-Id: I990b26056477e3e41c6afe82479383d8bc7c45d0
Reviewed-by: Jarek Kobus <jaroslaw.kobus@qt.io>
Reviewed-by: hjk <hjk@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>