Fix saving of keyboard schemes in presence of broken shortcut
QKeySequence can consist of weird keys, and QmlDesigner accidentally set
one to \u0002. Trying to write this into an XML attribute results in
QXmlStreamWriter simply failing and never writing anything.

Aside from fixing QmlDesigner in another patch, make sure that we only
write valid characters to the XML, by using the same checks as in
QXmlStreamWriter and writing/reading as hex-encoded as a fallback.

Also print a warning, if saving the keyboard schemes fails for any
reason.

Fixes: QTCREATORBUG-29431
Change-Id: Ief656ea42db3c1dceeb3f90a851eb9000b63f0c6
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>
Reviewed-by: Christian Stenger <christian.stenger@qt.io>