CMake/macOS: Fix framework paths with CMake >= 3.27
Previously the "compileGroups" > "includes" section contained paths to
frameworks, like "<Qt>/lib/QtCore.framework" which we use to add a
framework path of "<Qt>/lib" to the code model. These paths moved to a
new "compileGroups" > "frameworks" section in file-api in CMake >= 3.27,
so we need to read this section as well.

Fixes: QTCREATORBUG-29450
Change-Id: Ica685006d1373fdc3efee6029acf4a7f7e8528f3
Reviewed-by: Cristian Adam <cristian.adam@qt.io>
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>