RunControl: Revert reject button's text to keep apps running
Previously, when a user attempted to switch sessions or quit Qt Creator,
a dialog would appear asking whether to continue running the current app.
The options were:
- Force Quit
- Keep Running

However, in Qt Creator 10, the 'Keep Running' button was mislabelled as 'Cancel',
which does not accurately represent its function.
This commit corrects this issue by reverting the button's text back to 'Keep Running'.

Change-Id: Ic48648baf6b166907d986d58e827cb52adfd551c
Reviewed-by: Eike Ziller <eike.ziller@qt.io>
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>
Reviewed-by: Marcus Tillmanns <marcus.tillmanns@qt.io>