ProjectExplorer: prefer git ssh over windows ssh
In contrast to the git provided ssh the current ssh provided in windows
ignores the SSH_ASKPASS environment variable. So we should prefer the
implementation of git. This was done correctly in previous versions of
Qt Creator since Environment::searchInPath first went through the
additional provided directories before looking into the PATH environment
variable. This behavior changed in 11.0 and now we look first into the
entries of PATH.

Change-Id: I4f3c21866d2d1a5151998fd527978ed482946ef8
Reviewed-by: Marcus Tillmanns <marcus.tillmanns@qt.io>
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>