QmlDesigner: Use view instead of values
In the case of the has and remove functions we don't need the value. If
we use views we can use const expressions.

We have to change from QHash to std::map because QHash can not handle
different types for the find function.

The patch is removing a double lookup too.

Change-Id: I89bb942d7948127344d230ee61ac166102354ccb
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>
Reviewed-by: Qt CI Patch Build Bot <ci_patchbuild_bot@qt.io>
Reviewed-by: Vikas Pachdha <vikas.pachdha@qt.io>