Editor: transform the line column label into button
Having a button instead of a labal makes it more obvious that there is
an action triggered when clicking the label.

Also fix the flickering introduced by
f0bb7e8bfa9f9880a09872160c7765be1e44a594 by saving the maximum width of
the sizeHint. Since the width can get big if we have a selection we do
not save the size in that case to avoid blocking the width when the
selection is removed again.

Change-Id: If8210ecbabb8c6449c4f88652fbed99f8526057d
Reviewed-by: Eike Ziller <eike.ziller@qt.io>