QmlDesigner: Implement snapping to increment
When ctrl is pressed during move gizmo drag, the move is snapped to
configurable increment. If the drag axes are aligned to the global
axes, the snapping can be done either to absolute grid or increment
to start position, depending on snap options selected.

If drag axes are not aligned, then drag is simply snapped to
configured increment length along the drag vector regardless of
configuration.

All snapping settings are persistent and shared by all 3D scenes.

Fixes: QDS-10463
Change-Id: I9407e558aae0ad8b72d975fb6dfa15a4e78f6ee8
Reviewed-by: Mahmoud Badri <mahmoud.badri@qt.io>
Reviewed-by: Qt CI Patch Build Bot <ci_patchbuild_bot@qt.io>