Utils: Make SmallStringLiteral constexpr
The whole small string has to be intialized for const expressions. But
because we don't want to pay for that all the time we use a C++ feature
to only do that at compile time. So now at runtime only the control
block and the string content code is generated but at compile time
everything is filled up with zeros.

So you can now write:

void foo(std::vector<Utils::SmallString> &v) {
    static constexpr Utils::SmallStringLiteral foo{"Some short string"};

    v.push_back(foo);
}

The advantage compared to a string_view is that it is only memory
copying the literal because it know it has to be a constant.

Change-Id: Ia0e4bac293c363ae08341d18d47ddaf78e896e6a
Reviewed-by: Tim Jenssen <tim.jenssen@qt.io>
Reviewed-by: Qt CI Patch Build Bot <ci_patchbuild_bot@qt.io>