macOS: Allow apps be debugged via process stub
This fixes debugging with "Run in Terminal".
We usually do not use the disclaim tool when debugging,
because LLDB does that itself, but it looks like not having
get-task-allow prevents that in the first place.

Add that and disable-library-validation (which is needed for
get-task-allow when notarizing) to our process stub.

Fixes: QTCREATORBUG-29497
Change-Id: I86c184b2b108b62822bf89d80ea2b4ca9190f5b8
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>
Reviewed-by: Marcus Tillmanns <marcus.tillmanns@qt.io>