Revert "App: Replace QtLockedFile usage with QLockFile"
This reverts commit 1fca05959f0a321943d803941c2fcdb7f6f0258d.

Reason for revert: Portale +2'ed under the assumption that this
goes into the master branch. This year old issue should definitely
be fixed, but typically not that urgently in a .3 patch release

Let's revert this and move it on master, please. I take the blame for the messy git history.

Change-Id: I8ae5a5e8e89ce38e7fe663b09cb6c2c3ab520ab0
Reviewed-by: Cristian Adam <cristian.adam@qt.io>