Fix that target selector accumulated labels
for kit aspects. The widgets added to the "KitArea", including the
created labels for the kit aspects, are children of the widget itself,
not the layout, so since we do not delete the KitArea widget, we need to
explicitly delete the labels too when updating, the same as we do for
the aspects' widgets.

Fixes: QTCREATORBUG-29519
Change-Id: I605edd4969345d18064c1c00ec4d617499de8eb7
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>
Reviewed-by: hjk <hjk@qt.io>